requirements.txt
langchain
langchain-groq
langchain-core
langchainhub
python-dotenv
pydantic